Module08/requirements.txt
requests==2.32.3
chainlit==1.1.402
openai>=1.30.0
foundry-local-sdk
